Fixed ck mark & Confirm menu screen
I seem to be having that issue with the inventory menu not accessible until I click on something or specificly the dialogue volume bar that's triggering it somehow. I tried to figure out all visually on my end, but bar is the same as the music volume bar from what I can tell.